requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
aom=3.5.0=h63175ca_0
async-timeout=4.0.2=pypi_0
async_generator=1.10=py_0
attrs=22.2.0=pyh71513ae_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.0=pyha770c72_0
botocore=1.29.89=pypi_0
brotlipy=0.7.0=py310h8d17308_1005
bs4=4.12.0=hd8ed1ab_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2023.5.7=h56e8100_0
cairo=1.16.0=hdecc03f_1016
certifi=2022.12.7=pypi_0
cffi=1.15.1=py310h628cb3f_3
charset-normalizer=3.1.0=pypi_0
cryptography=40.0.1=py310h6e82f81_0
dav1d=1.2.1=hcfcfb64_0
docutils=0.20.1=py310h5588dad_0
exceptiongroup=1.1.1=pyhd8ed1ab_0
expat=2.5.0=h63175ca_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=h546665d_1
frozenlist=1.3.3=pypi_0
gettext=0.21.1=h5728263_0
glew=2.1.0=h39d44d4_2
gql=3.4.0=pypi_0
graphite2=1.3.13=1000
graphql-core=3.2.3=pypi_0
h11=0.14.0=pyhd8ed1ab_0
harfbuzz=7.3.0=h196d34a_0
icu=72.1=h63175ca_0
idna=3.4=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jmespath=1.0.1=pypi_0
kivy=2.2.1=py310h0aa0824_0
lcms2=2.15=h3e3b177_1
lerc=4.0.0=h63175ca_0
libavif=0.11.1=ha7b97ba_2
libdeflate=1.18=hcfcfb64_0
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libflac=1.4.3=h63175ca_0
libglib=2.76.4=he8f3873_0
libiconv=1.17=h8ffe710_0
libjpeg-turbo=2.1.5.1=hcfcfb64_0
libmad=0.15.1b=hcfcfb64_1001
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libsqlite=3.40.0=hcfcfb64_0
libtiff=4.5.1=h6c8260b_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.1=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libzlib=1.2.13=hcfcfb64_4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.2=py310h8d17308_0
mpg123=1.31.3=h63175ca_0
msys2-conda-epoch=20160418=1
multidict=6.0.4=pypi_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.1.1=hcfcfb64_1
outcome=1.2.0=pyhd8ed1ab_0
pcre2=10.40=h17e33f8_0
pillow=10.0.0=py310hb653ca7_0
pip=23.0.1=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
pthread-stubs=0.4=hcd874cb_1001
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.1.0=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.10.9=h4de0772_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.10=3_cp310
pyyaml=6.0=py310h8d17308_5
requests=2.28.2=pypi_0
requests-toolbelt=0.10.1=pypi_0
sdl2=2.26.5=h63175ca_0
sdl2_image=2.6.3=h3be12bc_1
sdl2_mixer=2.6.3=h090ff69_0
sdl2_ttf=2.20.2=h29dc827_1
selenium=4.8.2=pyhd8ed1ab_0
setuptools=67.6.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
sniffio=1.3.0=pyhd8ed1ab_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
trio=0.22.0=py310h5588dad_1
trio-websocket=0.10.2=pyhd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023b=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.14=pypi_0
vc=14.3=hb6edc58_10
vc14_runtime=14.36.32532=hfdfe4a8_17
vs2015_runtime=14.36.32532=h05e6639_17
websockets=10.4=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wsproto=1.2.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
yarl=1.8.2=pypi_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h12be248_7
